empty array ([]) for empty to-many relationships
According to [JSON
API](http://jsonapi.org/format/#document-resource-object-linkage) an emtpy array
should be returned for emtpy to-many relationships.

```
Resource linkage MUST be represented as one of the following:

null for empty to-one relationships.
an empty array ([]) for empty to-many relationships.
a single resource identifier object for non-empty to-one relationships.
an array of resource identifier objects for non-empty to-many relationships.
```